Fix UI problem 100
Fixed the bug now
